requirements.txt
django-google-maps
mysqlclient
pillow
django-tinymce
django-filer
django-rest-framework
markdown
django-filter